:sparkles: feat: update entity
- 엔티티 위치 변경
- 상태 추가
