+++ src/main/java/com/bookbook/booklink/borrow_service/model/BorrowStatus.java
@@ -6,7 +6,9 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public enum BorrowStatus {
-    BORROWED("대출 중"),
+    REQUESTED("대여 대기"),
+    BORROWED("대여 중"),
+    SUSPENDED("대여 중단"),
     EXTENDED("연장 중"),
     OVERDUE("연체 중"),
     RETURNED("반납 완료")
